feat(module): Slicing form edit modul instructor
